Move Posts listing back to / and simplify routing
- Restore / as Posts listing (original behavior)
- Remove redundant /posts listing page
- Update Nav Posts link to point to /
